+++ run_examples_mstatspop.sh
@@ -1,46 +1,17 @@
-#Validation: test a number of combinations.
-#-f fasta tfa ms
-#-i ./100Kchr10.fa
-#-i ./100Kchr10.tfa
-#-i ./100Kchr10_fa2ms_03.ms.txt
-#-g ./100Kchr10.gtf nonsynonymous Nuclear_Universal
-#-W ./coord_100Kb.txt 
-#-E ./100Kchr10_fa2tfa_02.tfa_nonsynonymous_max_IncludeMissing_NOoutg_ploidy1_WEIGHTS.txt
-#-o 1
-#-N 2 40 2 
-#-N 3 20 20 2
-#options: [-G -u  -T]
-#-G 1/0 -u 1/0
-
-#zlib 1.3 installation (dependency)
-#
-#mkdir -p ./zlib
-#wget http://zlib.net/zlib-1.3.tar.gz -P ./zlib
-#tar -zxvf ./zlib/zlib-1.3.tar.gz -C ./zlib
-#rm ./zlib/zlib-1.3.tar.gz
-#cd ./zlib/zlib-1.3
-#./configure
-#make
-#sudo make install
-
-#gsl installation (dependency)
-#mkdir -p /tmp/gsl && \
-#    curl -o /tmp/gsl-2.2.tar.gz ftp://ftp.gnu.org/gnu/gsl/gsl-2.2.tar.gz -LOk && \
-#    tar -zxvf /tmp/gsl-2.2.tar.gz -C /tmp/gsl && \
-#    rm /tmp/gsl-2.2.tar.gz && \
-#    cd /tmp/gsl/gsl-2.2 && \
-#    ./configure && \
-#    make && \
-#    sudo make install
-
 #To compile:
-#gcc ./sources/*.c -lm -o ./bin/mstatspop -Wall -O3 -lz
-#OR (IN CASE USING OPTIMAL TESTS include the GSL Scientific library, downloading from http://www.gnu.org/software/gsl/)
-#for osx 10.12 add:-I/usr/local/include -L/usr/local/lib
-gcc ./sources/*.c -lgsl -lgslcblas -lm -o ./bin/mstatspop -Wall -DinGSL=1 -O3 -L/opt/homebrew/Cellar/gsl/2.7.1/lib/ /opt/homebrew/Cellar/gsl/2.7.1/lib/libgsl.a -I/opt/homebrew/Cellar/gsl/2.7.1/include -lz
+#for osx: look at library versions and modify
+brew install gsl
+brew install htslib
+brew install zlib
+sh build.sh
+#OR
+#gcc  -lgsl -lgslcblas -lm -lhts -lz -o ./bin/mstatspop -Wall -DinGSL=1 -O3 ./mstatspop/zutil.c ./mstatspop/files_util.c ./mstatspop/freq_stats.c ./mstatspop/fstcalc.c ./mstatspop/get_msdata.c ./mstatspop/get_obsdata.c ./mstatspop/get_obsdatastats.c ./mstatspop/get_tfadata.c ./mstatspop/jointfreqdist.c ./mstatspop/log.c ./mstatspop/mismatch.c ./mstatspop/missing_freqs.c ./mstatspop/mstatspop.c ./mstatspop/optimal-RinfR0.c ./mstatspop/permute.c ./mstatspop/print_output.c ./mstatspop/ran1.c ./mstatspop/sancestral.c ./mstatspop/tfasta.c ./mstatspop/tn93.c ./mstatspop/usegff.c ./mstatspop/util.c ./mstatspop/zindex.c ./mstatspop/calc_Toptimal_tests.c ./mstatspop/calcFs.c ./mstatspop/calcR2.c -I/opt/homebrew/Cellar/gsl/2.8/include  -I/opt/homebrew/Cellar/htslib/1.21/include -L/opt/homebrew/Cellar/gsl/2.8/lib -L/opt/homebrew/Cellar/htslib/1.21/lib
 #for linux:
-#gcc ./sources/*.c -lgsl -lgslcblas -lm -o ./bin/mstatspop -Wall -DinGSL=1 -O3 -lz
-gcc -o ./bin/ms ./sources_msHudson/ms.c ./sources_msHudson/streec.c ./sources_msHudson/rand2.c -lm -O3
+#sh build.sh
+#gcc  -lgsl -lgslcblas -lm -lhts -lz -o ./bin/mstatspop -Wall -DinGSL=1 -O3 ./mstatspop/zutil.c ./mstatspop/files_util.c ./mstatspop/freq_stats.c ./mstatspop/fstcalc.c ./mstatspop/get_msdata.c ./mstatspop/get_obsdata.c ./mstatspop/get_obsdatastats.c ./mstatspop/get_tfadata.c ./mstatspop/jointfreqdist.c ./mstatspop/log.c ./mstatspop/mismatch.c ./mstatspop/missing_freqs.c ./mstatspop/mstatspop.c ./mstatspop/optimal-RinfR0.c ./mstatspop/permute.c ./mstatspop/print_output.c ./mstatspop/ran1.c ./mstatspop/sancestral.c ./mstatspop/tfasta.c ./mstatspop/tn93.c ./mstatspop/usegff.c ./mstatspop/util.c ./mstatspop/zindex.c ./mstatspop/calc_Toptimal_tests.c ./mstatspop/calcFs.c ./mstatspop/calcR2.c
+#to compile Hudson's ms
+#gcc -o ./bin/ms ./sources_msHudson/ms.c ./sources_msHudson/streec.c ./sources_msHudson/rand2.c -lm -O3
+
 cd ./Examples
 
 echo
@@ -83,10 +54,10 @@ echo Example fa.07.txt: Two pops + outg + missing + GTF2 for nonsyn
 echo ../bin/mstatspop  -f fasta -i ./100Kchr10.fa -o 0 -N 2 40 2 -T ../Results/mstatspop_100chr10.fa.07.txt -G 1 -u 1 -g ./100Kchr10.gtf nonsynonymous Nuclear_Universal -c max -K 1 -n chr10.txt
 ../bin/mstatspop  -f fasta -i ./100Kchr10.fa -o 0 -N 2 40 2 -T ../Results/mstatspop_100chr10.fa.07.txt -G 1 -u 1 -g ./100Kchr10.gtf nonsynonymous Nuclear_Universal -c max -K 1 -n chr10.txt
 echo 
-echo Example fa.08.txt: Three pops + outg + missing + permutation test
-echo ../bin/mstatspop  -f fasta -i ./100Kchr10.fa -o 0 -N 3 20 20 2 -T ../Results/mstatspop_100chr10.fa.08.txt -G 1 -u 1 -t 1000 -s 1684 -K 1 -n chr10.txt
-../bin/mstatspop  -f fasta -i ./100Kchr10.fa -o 0 -N 3 20 20 2 -T ../Results/mstatspop_100chr10.fa.08.txt -G 1 -u 1 -t 1000 -s 1684 -K 1 -n chr10.txt
-echo 
+echo Example fa.08.txt: Three pops + outg + permutation test
+echo ../bin/mstatspop  -f fasta -i ./100Kchr10.fa -o 0 -N 3 20 20 2 -T ../Results/mstatspop_100chr10.fa.08.txt -G 1 -u 0 -t 1000 -s 1684 -K 1 -n chr10.txt
+../bin/mstatspop  -f fasta -i ./100Kchr10.fa -o 0 -N 3 20 20 2 -T ../Results/mstatspop_100chr10.fa.08.txt -G 1 -u 0 -t 1000 -s 1684 -K 1 -n chr10.txt
+echo
 echo Example fa.09.txt: Three pops + outg + missing + reordering the two pops in inverse order
 echo ../bin/mstatspop  -f fasta -i ./100Kchr10.fa -o 0 -N 3 20 20 2 -T ../Results/mstatspop_100chr10.fa.09.txt -G 1 -K 1-u 1 -O 42 20 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 40 41 -n chr10.txt
 ../bin/mstatspop  -f fasta -i ./100Kchr10.fa -o 0 -N 3 20 20 2 -T ../Results/mstatspop_100chr10.fa.09.txt -G 1 -u 1 -K 1 -O 42 20 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 40 41 -n chr10.txt
@@ -160,10 +131,10 @@ echo Example ms.12.txt: Two pops + outg + missing + weight for nonsyn + permutat
 echo ../bin/mstatspop  -f ms -i ./100Kchr10_simulation.ms.txt -o 0 -N 2 40 2 -G 1 -T ../Results/mstatspop_100chr10.ms.12.txt -l 100000 -G 1 -u 1 -m ./100Kchr10_fa2ms_04b.ms.txt_npops2_nsam42_nonsynonymous_max_ExcludeMissingVariantsmhits_outg_ploidy1_MASK.txt -t 1000 -s 1684 -v 2.0 -n chr10.txt
 ../bin/mstatspop  -f ms -i ./100Kchr10_simulation.ms.txt -o 0 -N 2 40 2 -G 1 -T ../Results/mstatspop_100chr10.ms.12.txt -l 100000 -G 1 -u 1 -m ./100Kchr10_fa2ms_04b.ms.txt_npops2_nsam42_nonsynonymous_max_ExcludeMissingVariantsmhits_outg_ploidy1_MASK.txt -t 1000 -s 1684 -v 2.0 -n chr10.txt
 echo
-echo Example ms.13.txt: Two pops + outg + missing + weight for nonsyn + permutation test + s/v ratio + 1000 replicates
-../bin/ms 42 1000 -t 1000 -I 2 40 2 -ej 2.0 2 1 -p 9 > ./100Kchr10_simulation1000.ms.txt
-echo ../bin/mstatspop  -f ms -i ./100Kchr10_simulation1000.ms.txt -o 1 -N 2 40 2 -G 1 -T ../Results/mstatspop_100chr10.ms.13.txt -l 100000 -G 1 -u 1 -m ./100Kchr10_fa2ms_04b.ms.txt_npops2_nsam42_nonsynonymous_max_ExcludeMissingVariantsmhits_outg_ploidy1_MASK.txt -t 1000 -s 1684 -v 2.0 -r 1000 -n chr10.txt
-../bin/mstatspop  -f ms -i ./100Kchr10_simulation1000.ms.txt -o 1 -N 2 40 2 -G 1 -T ../Results/mstatspop_100chr10.ms.13.txt -l 100000 -G 1 -u 1 -m ./100Kchr10_fa2ms_04b.ms.txt_npops2_nsam42_nonsynonymous_max_ExcludeMissingVariantsmhits_outg_ploidy1_MASK.txt -t 1000 -s 1684 -v 2.0 -r 1000 -n chr10.txt
+echo Example ms.13.txt: Two pops + outg + missing + weight for nonsyn + permutation test + s/v ratio + 10 replicates
+../bin/ms 42 1000 -t 10 -I 2 40 2 -ej 2.0 2 1 -p 9 > ./100Kchr10_simulation10.ms.txt
+echo ../bin/mstatspop  -f ms -i ./100Kchr10_simulation10.ms.txt -o 1 -N 2 40 2 -G 1 -T ../Results/mstatspop_100chr10.ms.13.txt -l 100000 -G 1 -u 1 -m ./100Kchr10_fa2ms_04b.ms.txt_npops2_nsam42_nonsynonymous_max_ExcludeMissingVariantsmhits_outg_ploidy1_MASK.txt -s 1684 -v 2.0 -r 00 -n chr10.txt
+../bin/mstatspop  -f ms -i ./100Kchr10_simulation10.ms.txt -o 1 -N 2 40 2 -G 1 -T ../Results/mstatspop_100chr10.ms.13.txt -l 100000 -G 1 -u 1 -m ./100Kchr10_fa2ms_04b.ms.txt_npops2_nsam42_nonsynonymous_max_ExcludeMissingVariantsmhits_outg_ploidy1_MASK.txt -s 1684 -v 2.0 -r 10 -n chr10.txt
 echo
 echo --------------------------------------------------------------------------------------------------
 echo PARAMETERS FOR TFASTA INPUT -f tfa: -w [-z -Y -W -E -O  -T -s] SLIDING WINDOW ANALYSIS OF EMPIRICAL DATA
@@ -185,10 +156,10 @@ echo Example tfa.02.txt: Two populations
 echo ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 0 -N 2 40 2 -T ../Results/mstatspop_100chr10.tfa.02.txt -w 100000 -n chr10.txt
 ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 0 -N 2 40 2 -T ../Results/mstatspop_100chr10.tfa.02.txt -w 100000 -n chr10.txt
 echo 
-echo Example tfa.03.txt: Two pops + outgroup
-echo ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 0 -N 2 40 2 -T ../Results/mstatspop_100chr10.tfa.03.txt -G 1 -w 100000 -n chr10.txt
-../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 0 -N 2 40 2 -T ../Results/mstatspop_100chr10.tfa.03.txt -G 1 -w 100000 -n chr10.txt
-echo 
+echo Example tfa.03.txt: Three pops + outgroup
+echo ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 10 -N 3 20 20 2 -T ../Results/mstatspop_100chr10.tfa.03.txt -G 1 -w 100000 -n chr10.txt
+../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 10 -N 3 20 20 2 -T ../Results/mstatspop_100chr10.tfa.03.txt -G 1 -w 100000 -n chr10.txt
+echo
 echo Example tfa.04.txt: Two pops + outg + missing
 echo ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 0 -N 2 40 2 -T ../Results/mstatspop_100chr10.tfa.04.txt -G 1 -u 1 -w 100000 -n chr10.txt
 ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 0 -N 2 40 2 -T ../Results/mstatspop_100chr10.tfa.04.txt -G 1 -u 1 -w 100000 -n chr10.txt
@@ -249,21 +220,25 @@ echo Example tfa.15.txt: Two pops + outg + missing + window 10000 + slide 20000 
 echo ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 2 40 2 -T ../Results/mstatspop_100chr10.tfa.15.txt -G 1 -u 1 -w 10000 -z 20000 -t 1000 -s 1684 -n chr10.txt
  ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 2 40 2 -T ../Results/mstatspop_100chr10.tfa.15.txt -G 1 -u 1 -w 10000 -z 20000 -t 1000 -s 1684 -n chr10.txt
 echo
-echo Example tfa.16.txt: Two pops + outg + missing + window 1000 + slide 100 + permutations:
-echo ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 2 40 2 -T ../Results/mstatspop_100chr10.tfa.16.txt -G 1 -u 1 -w 1000 -z 100 -t 1000 -s 1684 -n chr10.txt
- ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 2 40 2 -T ../Results/mstatspop_100chr10.tfa.16.txt -G 1 -u 1 -w 1000 -z 100 -t 1000 -s 1684 -n chr10.txt
+echo Example tfa.16.txt: Two pops + outg + missing + window 100 + slide 100 + permutations:
+echo ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 2 40 2 -T ../Results/mstatspop_100chr10.tfa.16.txt -G 1 -u 1 -w 100 -z 100 -t 1000 -s 1684 -n chr10.txt
+ ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 2 40 2 -T ../Results/mstatspop_100chr10.tfa.16.txt -G 1 -u 1 -w 100 -z 100 -t 1000 -s 1684 -n chr10.txt
+echo 
+echo Example tfa.17.txt: Three pops + outg + missing + window 100 + slide 100 + permutations:
+echo ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 3 20 20 2 -T ../Results/mstatspop_100chr10.tfa.17.txt -G 1 -u 1 -w 100 -z 100 -t 1000 -s 1684 -n chr10.txt
+ ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 3 20 20 2 -T ../Results/mstatspop_100chr10.tfa.17.txt -G 1 -u 1 -w 100 -z 100 -t 1000 -s 1684 -n chr10.txt
 echo
-echo Example tfa.17.txt: Three pops + outg + missing + window 1000 + slide 100 + permutations:
-echo ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 3 20 20 2 -T ../Results/mstatspop_100chr10.tfa.17.txt -G 1 -u 1 -w 1000 -z 100 -t 1000 -s 1684 -n chr10.txt
- ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 3 20 20 2 -T ../Results/mstatspop_100chr10.tfa.17.txt -G 1 -u 1 -w 1000 -z 100 -t 1000 -s 1684 -n chr10.txt
+echo Example tfa.18.txt: Three pops LESS SAMPLES + outg + missing + window 100 + slide 100 :
+echo ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 3 10 10 2 -T ../Results/mstatspop_100chr10.tfa.18.txt -G 1 -u 1 -w 100 -z 100 -n chr10.txt
+ ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 3 10 10 2 -T ../Results/mstatspop_100chr10.tfa.18.txt -G 1 -u 1 -w 100 -z 100 -n chr10.txt
+echo 
+echo Example tfa.19.txt: Three pops LESS SAMPLES + reordering + outg + window 100 + slide 100 :
+echo ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 3 10 10 2 -T ../Results/mstatspop_100chr10.tfa.19a.txt -G 1 -u 0 -w 100 -z 100 -O 22 20 21 22 23 24 25 26 27 28 29 0 1 2 3 4 5 6 7 8 9 40 41 -n chr10.txt
+ ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 3 10 10 2 -T ../Results/mstatspop_100chr10.tfa.19a.txt -G 1 -u 0 -w 100 -z 100 -O 22 20 21 22 23 24 25 26 27 28 29 0 1 2 3 4 5 6 7 8 9 40 41 -n chr10.txt
 echo
-echo Example tfa.18.txt: Three pops LESS SAMPLES + outg + missing + window 1000 + slide 100 :
-echo ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 3 10 10 2 -T ../Results/mstatspop_100chr10.tfa.18.txt -G 1 -u 1 -w 1000 -z 100 -n chr10.txt
- ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 3 10 10 2 -T ../Results/mstatspop_100chr10.tfa.18.txt -G 1 -u 1 -w 1000 -z 100 -n chr10.txt
-echo
-echo Example tfa.19.txt: Three pops LESS SAMPLES + reordering + outg + missing + window 1000 + slide 100 :
-echo ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 3 10 10 2 -T ../Results/mstatspop_100chr10.tfa.19.txt -G 1 -u 1 -w 1000 -z 100 -O 22 20 21 22 23 24 25 26 27 28 29 0 1 2 3 4 5 6 7 8 9 40 41 -n chr10.txt
- ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 3 10 10 2 -T ../Results/mstatspop_100chr10.tfa.19.txt -G 1 -u 1 -w 1000 -z 100 -O 22 20 21 22 23 24 25 26 27 28 29 0 1 2 3 4 5 6 7 8 9 40 41 -n chr10.txt
+echo Example tfa.19.txt: Three pops LESS SAMPLES + reordering + outg + missing + window 100 + slide 100 :
+echo ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 3 10 10 2 -T ../Results/mstatspop_100chr10.tfa.19.txt -G 1 -u 1 -w 100 -z 100 -O 22 20 21 22 23 24 25 26 27 28 29 0 1 2 3 4 5 6 7 8 9 40 41 -n chr10.txt
+ ../bin/mstatspop -f tfa -i ./100Kchr10.tfa.gz -o 1 -N 3 10 10 2 -T ../Results/mstatspop_100chr10.tfa.19.txt -G 1 -u 1 -w 100 -z 100 -O 22 20 21 22 23 24 25 26 27 28 29 0 1 2 3 4 5 6 7 8 9 40 41 -n chr10.txt
 echo
 echo --------------------------------------------------------------------------------------------------
 echo Examples using MULTICHROMOSOME TFA FILE
@@ -300,6 +275,14 @@ echo
 echo Same but in single line output
 echo ../bin/mstatspop -f tfa -i ./100Kallchr.tfa.gz -o 1 -N 2 40 2 -G 1 -T ../Results/mstatspop_100allchr.tfa.06.txt  -u 1 -w 10000 -n chr10-12-14.txt
 ../bin/mstatspop -f tfa -i ./100Kallchr.tfa.gz -o 1 -N 2 40 2 -G 1 -T ../Results/mstatspop_100allchr.tfa.06.txt  -u 1 -w 10000 -n chr10-12-14.txt
+echo
+echo Same but in single line output, no missing and three pops
+echo ../bin/mstatspop -f tfa -i ./100Kallchr.tfa.gz -o 1 -N 3 20 20 2 -G 1 -T ../Results/mstatspop_100allchr.tfa.06b.txt  -u 0 -w 10000 -n chr10-12-14.txt
+../bin/mstatspop -f tfa -i ./100Kallchr.tfa.gz -o 1 -N 3 20 20 2 -G 1 -T ../Results/mstatspop_100allchr.tfa.06b.txt  -u 0 -w 10000 -n chr10-12-14.txt
+echo
+echo Same but in single line output, no missing and three pops and NO outgroup
+echo ../bin/mstatspop -f tfa -i ./100Kallchr.tfa.gz -o 1 -N 3 20 20 2 -G 0 -T ../Results/mstatspop_100allchr.tfa.06c.txt  -u 0 -w 10000 -n chr10-12-14.txt
+../bin/mstatspop -f tfa -i ./100Kallchr.tfa.gz -o 1 -N 3 20 20 2 -G 0 -T ../Results/mstatspop_100allchr.tfa.06c.txt  -u 0 -w 10000 -n chr10-12-14.txt
 echo
 echo Same but using displacement of 20000bp
 echo ../bin/mstatspop -f tfa -i ./100Kallchr.tfa.gz -o 1 -N 2 40 2 -G 1 -T ../Results/mstatspop_100allchr.tfa.07.txt  -u 1 -z 20000 -w 10000 -n chr10-12-14.txt
@@ -341,4 +324,48 @@ echo  ../bin/mstatspop -f fasta -i ./MC1R_PigsOutg_aligned.fas -o 0 -p 1 -u 0 -t
  echo ../bin/mstatspop -f fasta -i ./MC1R_PigsOutg_aligned.fas -o 0 -p 1 -u 0 -s 123456 -G 1 -N 3 48 46 1 -A ./MC1R_H1frq.txt -S ./MC1R_H0frq.txt -T ./MC1R_PigsOutg_NSyn_Opttest_H0.txt -n MC1R.txt
   ../bin/mstatspop -f fasta -i ./MC1R_PigsOutg_aligned.fas -o 0 -p 1 -u 0 -s 123456 -G 1 -N 3 48 46 1 -A ./MC1R_H1frq.txt -S ./MC1R_H0frq.txt -T ./MC1R_PigsOutg_NSyn_Opttest_H0.txt -n MC1R.txt
 echo
+echo --------------------------------------------------------------------------------------------------
+echo "indexing tfasta files:"
+echo --------------------------------------------------------------------------------------------------
+echo
+echo #Usage: ./tfa_index  [options] <input.tfa|input.tfa.bgz|input.tfa.gz|weights.txt|weights.txt.gz>
+echo #Options:
+echo # --version
+echo # --help
+echo # --threads <int>
+echo # --force      "force to overwrite the file"
+echo # --weight     "the input file is a weight file. In this case create and index for it. and convert it to the correct format"
+echo # --output FILE "set custom name to the output file, only used when converting from TFAv1 to TFAv2 or compressing the input file"
+echo # --min-shift <int> "set minimal interval size for CSI indices to 2^INT recommended value is 14, 0 means build tbi index. Default [0]"
+echo
+echo "Converting files to tfa.gz v2"
+echo # list of files
+# Examples/V0.1.0/100Kallchr07.tfa.gz_npops2_nsam42_nonsynonymous_max_IncludeMissingVariantsmhits_outg_ploidy1_WEIGHTS.gz
+echo "Converting Examples/V0.1.0/100Kallchr07.tfa.gz_npops2_nsam42_nonsynonymous_max_IncludeMissingVariantsmhits_outg_ploidy1_WEIGHTS.gz"
+../bin/tfa_index  ./V0.1.0/100Kallchr07.tfa.gz_npops2_nsam42_nonsynonymous_max_IncludeMissingVariantsmhits_outg_ploidy1_WEIGHTS.gz -f -o ../Results/V1.0.0/100Kallchr07.tfa.gz_npops2_nsam42_nonsynonymous_max_IncludeMissingVariantsmhits_outg_ploidy1_WEIGHTS.gz
+echo # Examples/V0.1.0/100Kallchr.tfa.gz
+echo "Converting Examples/V0.1.0/100Kallchr.tfa.gz"
+../bin/tfa_index  ./V0.1.0/100Kallchr.tfa.gz -f -o ../Results/V1.0.0/100Kallchr.tfa.gz
+echo
+echo # Examples/V0.1.0/100Kchr10_fa2tfa_08.tfa.gz_npops2_nsam42_nonsynonymous_max_IncludeMissingVariantsmhits_outg_ploidy1_WEIGHTS.gz
+echo "Converting Examples/V0.1.0/100Kchr10_fa2tfa_08.tfa.gz_npops2_nsam42_nonsynonymous_max_IncludeMissingVariantsmhits_outg_ploidy1_WEIGHTS.gz"
+../bin/tfa_index  ./V0.1.0/100Kchr10_fa2tfa_08.tfa.gz_npops2_nsam42_nonsynonymous_max_IncludeMissingVariantsmhits_outg_ploidy1_WEIGHTS.gz -f -o ../Results/V1.0.0/100Kchr10_fa2tfa_08.tfa.gz_npops2_nsam42_nonsynonymous_max_IncludeMissingVariantsmhits_outg_ploidy1_WEIGHTS.gz
+echo # Examples/V0.1.0/100Kchr10_tfa2tfa_03c.tfa.gz_npops3_nsam42_nonsynonymous_max_IncludeMissingVariantsmhits_outg_ploidy1_WEIGHTS.gz
+echo "Converting Examples/V0.1.0/100Kchr10_tfa2tfa_03c.tfa.gz_npops3_nsam42_nonsynonymous_max_IncludeMissingVariantsmhits_outg_ploidy1_WEIGHTS.gz"
+../bin/tfa_index  ./V0.1.0/100Kchr10_tfa2tfa_03c.tfa.gz_npops3_nsam42_nonsynonymous_max_IncludeMissingVariantsmhits_outg_ploidy1_WEIGHTS.gz -f -o ../Results/V1.0.0/100Kchr10_tfa2tfa_03c.tfa.gz_npops3_nsam42_nonsynonymous_max_IncludeMissingVariantsmhits_outg_ploidy1_WEIGHTS.gz
+echo # Examples/V0.1.0/100Kchr10.tfa.gz
+echo "Converting Examples/V0.1.0/100Kchr10.tfa.gz"
+../bin/tfa_index  ./V0.1.0/100Kchr10.tfa.gz -f -o ../Results/V1.0.0//100Kchr10.tfa.gz
+echo
+echo --------------------------------------------------------------------------------------------------
+echo "Merge tfasta files (same assembly reference but different individuals):"
+echo --------------------------------------------------------------------------------------------------
+echo
+echo #Usage: ./tfa_merge [options]
+echo #Options:
+echo # -i, --input FILE      Input file name (can be specified multiple times)
+echo # -o, --output FILE     Output file name
+echo # -f, --force           Force overwrite of output file
+echo # -h, --help            This help message
+echo #     --version         Print version
 echo